Enable forward navigation with pages
Add variables for tracking the current position, if this is the last page of the chapter and the total elements in that chapter.
Add LoadChapter function for loading the html of the current chapter and update the LoadPage function to update the variables when calling AppendChildren.
Also update the AppendChildren function with root to prevent adding the body element a second time. Also check for empty text nodes and if the readerPosition is smaller than currentPosition, don't add the elements so that the elements of the previous pages are not rendered.