Rewrite the html rendering to save the pages in an array
Enable forward and backward navigation of pages by saving the html of each page in an array. When changing the chapter, create the html of the chapter if it wasn't created and then show the content of the current page.
Also remove the functions PrependChildren, LoadChapter and RenderHtml and add new functions for generating the html nad for showing it.