Create missing pages when resizing and going to the previous page
Rename HtmlPage.position to startPosition and add endPosition to keep track of the elements that are in the html of the page.
In RerenderCurrentPage check if there is a gap between the current page and the next page, which occurs when making the window smaller and going to the previous page. If this is the case, add a new page between the current and the next page with the missing elements by calling RerenderCurrentPage recursively.