Completely rewrite BookContentComponent to render the html on hidden element
Change the layout to render the html of the book in a div hidden behind the iframe, which shows the prepared html of the book. Add chapter and position to Book class. Also simplify the functions in BookContentComponent, add Utils class for general functions and add getter functions in BookChapter.