Rewrite BookContentComponent to render the entire chapter in iframe and scroll to the correct position
Remove all unnecessary functions, variables, classes and elements. Replace the render div with the iframe and bottom curtain div for rendering the html and hiding the left over html of the page.
Also remove BookChapter class and functions for rendering each element and measuring the height. Instead render the entire chapter in the iframe and go through each element in the iframe to find the positions for each page. Then when navigating to another page, scroll to the appropriate position.