Use two iframes for ddisplaying the ebook content
Use one iframe which is on top and shows the content. When changing the page, render the new content on the hidden iframe and switch the places of the iframes, so that when the hidden iframe has rendered the content, it will be the iframe that is visible to the user.