Add third viewer and rewrite ShowPage for three viewers
When moving forward in the book, move the viewers clockwise and render the next page on the next page viewer
When moving backward in the book, move the viewers counterclockwise and render the previous page on the previous page viewer.
Add explanation and documentation for this behaviour in ShowPage and add smaller methods for breaking down the logic of ShowPage.